环境管理增加删除环境和复制环境功能

+++ internal/server/modules/repo/environment.go
@@ -250,9 +250,9 @@ func (r *EnvironmentRepo) ListVariableByProject(projectId uint) (vars []model.En
 	return
 }
 
-func (r *EnvironmentRepo) SaveEnvironment(environment model.Environment) (err error) {
+func (r *EnvironmentRepo) SaveEnvironment(environment *model.Environment) (err error) {
 	return r.DB.Transaction(func(tx *gorm.DB) error {
-		err = r.BaseRepo.Save(environment.ID, &environment)
+		err = r.BaseRepo.Save(environment.ID, environment)
 		if err != nil {
 			return err
 		}
@@ -268,12 +268,34 @@ func (r *EnvironmentRepo) SaveEnvironment(environment model.Environment) (err er
 	})
 }
 
+func (r *EnvironmentRepo) DeleteEnvironment(id uint) (err error) {
+	return r.DB.Transaction(func(tx *gorm.DB) error {
+		err = r.DB.Delete(&model.Environment{}, id).Error
+		if err != nil {
+			return err
+		}
+		err = r.DB.Delete(&model.ServeServer{}, "environment_id=?", id).Error
+		if err != nil {
+			return err
+		}
+		err = r.DB.Delete(&model.EnvironmentVar{}, "environment_id=?", id).Error
+		if err != nil {
+			return err
+		}
+		return nil
+	})
+}
+
 func (r *EnvironmentRepo) SaveVars(projectId, environmentId uint, environmentVars []model.EnvironmentVar) (err error) {
+	if len(environmentVars) == 0 {
+		return
+	}
 	err = r.DB.Delete(&model.EnvironmentVar{}, "environment_id=? and project_id=?", environmentId, projectId).Error
 	if err != nil {
 		return err
 	}
 	for key, _ := range environmentVars {
+		environmentVars[key].ID = 0
 		environmentVars[key].EnvironmentId = environmentId
 		environmentVars[key].ProjectId = projectId
 	}
@@ -290,21 +312,38 @@ func (r *EnvironmentRepo) GetListByProjectId(projectId uint) (environments []mod
 		return
 	}
 	for key, _ := range environments {
-		var vars []model.EnvironmentVar
-		err = r.DB.Find(&vars, "environment_id=?", environments[key].ID).Error
+		err = r.GetEnvironment(&environments[key])
 		if err != nil {
 			return
 		}
-		environments[key].Vars = vars
-
-		var servers []model.ServeServer
-		err = r.DB.Select("`biz_project_serve_server`.`id`,`biz_project_serve_server`.`created_at`,`biz_project_serve_server`.`updated_at`,`biz_project_serve_server`.`deleted`,`biz_project_serve_server`.`disabled`,`biz_project_serve_server`.`environment_id`,`biz_project_serve_server`.`serve_id`,`biz_project_serve_server`.`url`,`biz_project_serve_server`.`description`,biz_project_serve.name serve_name").Joins("left join biz_project_serve on biz_project_serve_server.serve_id = biz_project_serve.id").Find(&servers, "environment_id=?", environments[key].ID).Error
-		if err != nil {
-			return
-		}
-		environments[key].ServeServers = servers
-
 	}
+	return
+}
+
+func (r *EnvironmentRepo) GetEnvironmentById(id uint) (env *model.Environment, err error) {
+	err = r.DB.First(&env, id).Error
+	return
+}
+
+func (r *EnvironmentRepo) GetEnvironment(env *model.Environment) (err error) {
+	var vars []model.EnvironmentVar
+	err = r.DB.Find(&vars, "environment_id=?", env.ID).Error
+	if err != nil {
+		return
+	}
+	env.Vars = vars
+
+	var servers []model.ServeServer
+	err = r.DB.Find(&servers, "environment_id=?", env.ID).Error
+	if err != nil {
+		return
+	}
+	for key, server := range servers {
+		var serve model.Serve
+		r.DB.First(&serve, "id=?", server.ServeId)
+		servers[key].ServeName = serve.Name
+	}
+	env.ServeServers = servers
 	return
 }
